refactor: rename consumer
