+++ cmd/consumer/main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/kappusuton-yon-tebaru/backend/cmd/builder-consumer/builderconsumer"
+	"github.com/kappusuton-yon-tebaru/backend/cmd/consumer/consumer"
 	"github.com/kappusuton-yon-tebaru/backend/internal/utils"
 	"go.uber.org/zap"
 )
@@ -15,7 +15,7 @@ func main() {
 	app.Logger.Info("builder consumer initalizing")
 
 	app.Logger.Info("connecting to rmq", zap.String("queue_name", app.Config.BuilderConfig.QueueName), zap.String("queue_uri", app.Config.BuilderConfig.QueueUri))
-	msgs, err := app.RmqClient.Ch.Consume(app.Config.BuilderConfig.QueueName, "agent-builder-consumer", false, false, false, false, nil)
+	msgs, err := app.RmqClient.Ch.Consume(app.Config.BuilderConfig.QueueName, "agent-consumer", false, false, false, false, nil)
 	if err != nil {
 		panic(err)
 	}
